requirements.txt
flask
opencv-python
numpy
pandas
pillow
sqlalchemy
python-dotenv
gunicorn
flask_sqlalchemy